+++ sdk/azidentity/default_azure_credential.go
@@ -5,6 +5,7 @@ package azidentity
 
 import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/log"
 )
 
@@ -13,17 +14,21 @@ const (
 	developerSignOnClientID = "04b07795-8ddb-461a-bbee-02f9e1bf7b46"
 )
 
-// DefaultAzureCredentialOptions contains options for configuring how credentials are acquired.
+// DefaultAzureCredentialOptions contains options for configuring authentication. These options
+// may not apply to all credentials in the default chain.
 type DefaultAzureCredentialOptions struct {
-	// set this field to true in order to exclude the AzureCLICredential from the set of
-	// credentials that will be used to authenticate with
-	ExcludeAzureCLICredential bool
-	// set this field to true in order to exclude the EnvironmentCredential from the set of
-	// credentials that will be used to authenticate with
-	ExcludeEnvironmentCredential bool
-	// set this field to true in order to exclude the ManagedIdentityCredential from the set of
-	// credentials that will be used to authenticate with
-	ExcludeMSICredential bool
+	// The host of the Azure Active Directory authority. The default is AzurePublicCloud.
+	// Leave empty to allow overriding the value from the AZURE_AUTHORITY_HOST environment variable.
+	AuthorityHost AuthorityHost
+	// HTTPClient sets the transport for making HTTP requests
+	// Leave this as nil to use the default HTTP transport
+	HTTPClient policy.Transporter
+	// Retry configures the built-in retry policy behavior
+	Retry policy.RetryOptions
+	// Telemetry configures the built-in telemetry policy behavior
+	Telemetry policy.TelemetryOptions
+	// Logging configures the built-in logging policy behavior.
+	Logging policy.LogOptions
 }
 
 // NewDefaultAzureCredential provides a default ChainedTokenCredential configuration for applications that will be deployed to Azure.  The following credential
@@ -40,31 +45,33 @@ func NewDefaultAzureCredential(options *DefaultAzureCredentialOptions) (*Chained
 		options = &DefaultAzureCredentialOptions{}
 	}
 
-	if !options.ExcludeEnvironmentCredential {
-		envCred, err := NewEnvironmentCredential(nil)
-		if err == nil {
-			creds = append(creds, envCred)
-		} else {
-			errMsg += err.Error()
-		}
+	envCred, err := NewEnvironmentCredential(&EnvironmentCredentialOptions{AuthorityHost: options.AuthorityHost,
+		HTTPClient: options.HTTPClient,
+		Logging:    options.Logging,
+		Retry:      options.Retry,
+		Telemetry:  options.Telemetry,
+	})
+	if err == nil {
+		creds = append(creds, envCred)
+	} else {
+		errMsg += err.Error()
 	}
 
-	if !options.ExcludeMSICredential {
-		msiCred, err := NewManagedIdentityCredential("", nil)
-		if err == nil {
-			creds = append(creds, msiCred)
-		} else {
-			errMsg += err.Error()
-		}
+	msiCred, err := NewManagedIdentityCredential("", &ManagedIdentityCredentialOptions{HTTPClient: options.HTTPClient,
+		Logging:   options.Logging,
+		Telemetry: options.Telemetry,
+	})
+	if err == nil {
+		creds = append(creds, msiCred)
+	} else {
+		errMsg += err.Error()
 	}
 
-	if !options.ExcludeAzureCLICredential {
-		cliCred, err := NewAzureCLICredential(nil)
-		if err == nil {
-			creds = append(creds, cliCred)
-		} else {
-			errMsg += err.Error()
-		}
+	cliCred, err := NewAzureCLICredential(nil)
+	if err == nil {
+		creds = append(creds, cliCred)
+	} else {
+		errMsg += err.Error()
 	}
 
 	// if no credentials are added to the slice of TokenCredentials then return a CredentialUnavailableError